+++ src/contentscripts/ui/sidebar-host.js
@@ -10,12 +10,18 @@ listenToWindowEvent('member-change', async (ev) => {
 });
 
 async function initializeSidebar() {
-    const sidebarSrc = chrome.runtime.getURL('src/sidebar/sidebar.html');
+    const displayname = await getDisplayName(); // First set nickname if not done yet
+    const sidebarPage = displayname ? 'sidebar.html' : 'set-nickname.html';
+    const sidebarSrc = chrome.runtime.getURL('src/sidebar/' + sidebarPage);
     const sidebarClass = window.isOnAmazonWebsite ? '' : 'primevideo';
     jQuery('body')
         .css('display', 'flex')
         .append('' +
             '<div id="pvp-sidebar" class="' + sidebarClass + '">' +
+                '<div class="size-toggle">' +
+                    '<span class="collapse">►</span>' +
+                    '<span class="open">◄</span>' +
+                '</div>' +
             '<iframe id="pvp-sidebar-iframe" src="' + sidebarSrc + '"></iframe>' +
             '</div>');
     jQuery('#a-page').css('transition', 'width 1s ease-in-out').css('width', '100%');
@@ -38,6 +44,9 @@ async function initializeSidebar() {
         }
         alignSidebarWidth('15%');
     }, 5000); // Just to make sure it's not missing the 'start-video' message
+
+    // Watch & toggle collapse
+    handleCollapse();
 
     // Watcher is changing the theme
     startWebplayerWatcher();
@@ -68,14 +77,38 @@ function startWebplayerWatcher() {
     listenToWindowEvent('start-video', () => {
         const sidebar = jQuery('#pvp-sidebar');
         sidebar.addClass('player-mode');
-        alignSidebarWidth('15%');
     });
 
     listenToWindowEvent('close-video', () => {
         const sidebar = jQuery('#pvp-sidebar');
         sidebar.removeClass('player-mode');
-        alignSidebarWidth('15%');
     });
+}
+
+/**
+ * (Proactively) handle collapsing
+ * and opening the sidebar.
+ */
+function handleCollapse() {
+    const sidebar = jQuery('#pvp-sidebar');
+    const observer = new MutationObserver(() => {
+        alignSidebarWidth('15%'); // If collapsed, it will be size 0.
+    });
+    const observeConfig = {attributes: true, characterData: false};
+    observer.observe(sidebar[0], observeConfig);
+
+    jQuery('#pvp-sidebar .size-toggle').click(() => {
+        toggleSidebarCollapse();
+    });
+}
+
+function toggleSidebarCollapse() {
+    const sidebar = jQuery('#pvp-sidebar');
+    if (sidebar.hasClass('collapsed')) {
+        sidebar.removeClass('collapsed');
+    } else {
+        sidebar.addClass('collapsed');
+    }
 }
 
 /**
@@ -84,8 +117,10 @@ function startWebplayerWatcher() {
  * @param sidebarWidth like '300px' or '15%'
  */
 function alignSidebarWidth(sidebarWidth) {
+    const sidebar = jQuery('#pvp-sidebar');
+    if (sidebar.hasClass('collapsed')) { sidebarWidth = '0px'; } // If collapsed always enforce 0
     jQuery('#a-page').css('width', 'calc(100% - ' + sidebarWidth + ')');
-    jQuery('#pvp-sidebar').css('width', sidebarWidth);
+    sidebar.css('width', sidebarWidth);
 
     if (isLegacyWebPlayer()) {
         const webPlayer = jQuery('#dv-web-player');